\name{hpie}
\alias{Hurlbert's Probability of Interspecific Encounter}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Hurlbert's Probability of Interspecific Encounter
}
\description{
Calculates the probability of two individuals of different species to encounter each other randomly.
}
\usage{
hpie(data)
}
\arguments{
  \item{data}{
a vector of species abundances
}
}
\value{
Numeric, probability value
}
\references{
Hurlbert, S.H. (1971), The Nonconcept of Species Diversity: A Critique and Alternative Parameters. Ecology, 52: 577-586. https://doi.org/10.2307/1934145
}
